requirements.txt
gmpy2
mpmath
pycryptodome==3.20.0
pycryptodomex==3.20.0
sympy
